fix: refactor the table and split the api layer from the behaviour
* also introduce a context for the instance drawer
