+++ src/app/common/MASTable/CustomRowWrapper.tsx
@@ -1,16 +1,23 @@
-import React, { createContext, useContext } from 'react';
+import React, { createContext, LegacyRef, useContext } from 'react';
 import { InstanceStatus } from '@app/utils';
 import { css } from '@patternfly/react-styles';
 import './CustomRowWrapper.css';
+import { IRow } from '@patternfly/react-table';
 
-export type CustomRowWrapperContextProps = {
+export type CustomRowWrapperContextProps<T> = {
   activeRow?: string;
-  onRowClick?: (event: MouseEvent, rowIndex: number, row: any) => void;
+  onRowClick?: (
+    event: React.MouseEvent<T>,
+    rowIndex?: number,
+    row?: IRow
+  ) => void;
   rowDataTestId?: string;
   loggedInUser?: string;
 };
 
-const CustomRowWrapperContext = createContext<CustomRowWrapperContextProps>({
+const CustomRowWrapperContext = createContext<
+  CustomRowWrapperContextProps<any>
+>({
   activeRow: '',
   onRowClick: () => {
     // No-op
@@ -20,24 +27,25 @@ const CustomRowWrapperContext = createContext<CustomRowWrapperContextProps>({
 
 export const CustomRowWrapperProvider = CustomRowWrapperContext.Provider;
 
-export const CustomRowWrapper = (rowWrapperProps) => {
+export const CustomRowWrapper = (rowWrapperProps): JSX.Element => {
   const { activeRow, onRowClick, rowDataTestId, loggedInUser } = useContext(
     CustomRowWrapperContext
   );
   const { trRef, className, rowProps, row, ...props } = rowWrapperProps || {};
-  const { rowIndex } = rowProps;
-  const { isExpanded, originalData } = row;
   const isRowDeleted =
-    originalData?.status === InstanceStatus.DEPROVISION ||
-    originalData?.status === InstanceStatus.DELETED;
-  const isLoggedInUserOwner = loggedInUser === originalData?.owner;
+    row?.originalData?.status === InstanceStatus.DEPROVISION ||
+    row?.originalData?.status === InstanceStatus.DELETED;
+  const isLoggedInUserOwner = loggedInUser === row?.originalData?.owner;
   const isRowDisabled = isRowDeleted || !isLoggedInUserOwner;
+
+  const ref =
+    trRef === undefined ? undefined : (trRef as LegacyRef<HTMLTableRowElement>);
 
   return (
     <tr
       data-testid={rowDataTestId}
       tabIndex={!isRowDisabled ? 0 : undefined}
-      ref={trRef}
+      ref={ref}
       className={css(
         className,
         'pf-c-table-row__item',
@@ -46,12 +54,14 @@ export const CustomRowWrapper = (rowWrapperProps) => {
           : isLoggedInUserOwner && 'pf-m-selectable',
         !isRowDisabled &&
           activeRow &&
-          activeRow === originalData?.name &&
+          activeRow === row?.originalData?.name &&
           'pf-m-selected'
       )}
-      hidden={isExpanded !== undefined && !isExpanded}
-      onClick={(event: MouseEvent) =>
-        !isRowDisabled && onRowClick && onRowClick(event, rowIndex, row)
+      hidden={row?.isExpanded !== undefined && !row?.isExpanded}
+      onClick={(event) =>
+        !isRowDisabled &&
+        onRowClick &&
+        onRowClick(event, rowProps?.rowIndex, row)
       }
       {...props}
     />
