+++ src/app/modules/InstanceDrawer/InstanceDrawerContent.tsx
@@ -1,65 +1,60 @@
-import { DetailsTabProps } from '@app/modules/InstanceDrawer/DetailsTab';
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { InstanceStatus } from '@app/utils';
 import { MASLoading } from '@app/common';
 import { Tab, Tabs, TabTitleText } from '@patternfly/react-core';
 import { ConnectionTabProps } from '@app/modules/InstanceDrawer/ConnectionTab';
+import { useInstanceDrawer } from '@app/modules/InstanceDrawer/contexts/InstanceDrawerContext';
+import { InstanceDrawerTab } from '@app/modules/InstanceDrawer/tabs';
 
 export const ResourcesTab = React.lazy(() => import('./ConnectionTab'));
 export const DetailsTab = React.lazy(() => import('./DetailsTab'));
 
-export enum InstanceDrawerTabs {
-  DETAILS = 'details',
-  CONNECTION = 'connection',
-}
-
-export type InstanceDrawerContentProps = DetailsTabProps &
-  Pick<ConnectionTabProps, 'tokenEndPointUrl'> & {
-    initialTab?: InstanceDrawerTabs;
-  };
+export type InstanceDrawerContentProps = Pick<
+  ConnectionTabProps,
+  'tokenEndPointUrl'
+>;
 
 export const InstanceDrawerContent: React.FunctionComponent<InstanceDrawerContentProps> =
-  ({ initialTab, instanceDetail, tokenEndPointUrl }) => {
+  ({ tokenEndPointUrl }) => {
     const { t } = useTranslation();
 
-    const [activeTab, setActiveTab] = useState<InstanceDrawerTabs>(
-      initialTab || InstanceDrawerTabs.DETAILS
-    );
+    const { instanceDrawerTab, setInstanceDrawerTab, instanceDrawerInstance } =
+      useInstanceDrawer();
 
     const selectTab = (tab: string | number) => {
-      if (tab === InstanceDrawerTabs.CONNECTION) {
-        setActiveTab(InstanceDrawerTabs.CONNECTION);
+      if (tab === InstanceDrawerTab.CONNECTION) {
+        setInstanceDrawerTab(InstanceDrawerTab.CONNECTION);
       } else {
-        setActiveTab(InstanceDrawerTabs.DETAILS);
+        setInstanceDrawerTab(InstanceDrawerTab.DETAILS);
       }
     };
 
     const getExternalServer = () => {
-      const { bootstrap_server_host } = instanceDetail || {};
+      const { bootstrap_server_host } = instanceDrawerInstance || {};
       return bootstrap_server_host?.endsWith(':443')
         ? bootstrap_server_host
         : `${bootstrap_server_host}:443`;
     };
 
     const isKafkaPending =
-      instanceDetail?.status === InstanceStatus.ACCEPTED ||
-      instanceDetail?.status === InstanceStatus.PREPARING;
+      instanceDrawerInstance?.status === InstanceStatus.ACCEPTED ||
+      instanceDrawerInstance?.status === InstanceStatus.PREPARING;
 
     return (
       <React.Suspense fallback={<MASLoading />}>
         <Tabs
-          activeKey={activeTab.toString()}
+          activeKey={instanceDrawerTab.toString()}
           onSelect={(_, tab) => selectTab(tab)}
         >
           <Tab
-            eventKey={InstanceDrawerTabs.DETAILS.toString()}
+            eventKey={InstanceDrawerTab.DETAILS.toString()}
             title={<TabTitleText>{t('details')}</TabTitleText>}
           >
-            <DetailsTab instanceDetail={instanceDetail} />
+            <DetailsTab />
           </Tab>
           <Tab
-            eventKey={InstanceDrawerTabs.CONNECTION.toString()}
+            eventKey={InstanceDrawerTab.CONNECTION.toString()}
             title={<TabTitleText>{t('connection')}</TabTitleText>}
             data-testid='drawerStreams-tabConnect'
           >
