+++ src/app/modules/InstanceDrawer/tabs.ts
@@ -1,0 +1,4 @@
+export enum InstanceDrawerTab {
+  DETAILS = 'details',
+  CONNECTION = 'connection',
+}
