+++ src/app/modules/OpenshiftStreams/OpenshiftStreamsFederated.tsx
@@ -1,17 +1,19 @@
 import React from 'react';
 import { I18nextProvider } from 'react-i18next';
-import {
-  OpenshiftStreams,
-  OpenShiftStreamsProps,
-} from '@app/modules/OpenshiftStreams/OpenshiftStreams';
 import { PaginationProvider } from '@app/common';
 import { initI18N } from '@i18n/i18n';
 import { FederatedContext, FederatedProps } from '@app/contexts';
 import { ModalProvider } from '@rhoas/app-services-ui-components';
 import { KasModalLoader } from '@app/modals';
+import { InstanceDrawerContextProvider } from '@app/modules/InstanceDrawer/contexts/InstanceDrawerContext';
+import { KasLayout } from '@app/modules/OpenshiftStreams/components';
+import {
+  StreamsTableConnected,
+  StreamsTableProps,
+} from '@app/modules/OpenshiftStreams/components/StreamsTableConnected';
 
 // Version of OpenshiftStreams for federation
-type OpenshiftStreamsFederatedProps = OpenShiftStreamsProps & FederatedProps;
+type OpenshiftStreamsFederatedProps = StreamsTableProps & FederatedProps;
 
 const OpenshiftStreamsFederated: React.FunctionComponent<OpenshiftStreamsFederatedProps> =
   ({
@@ -33,13 +35,16 @@ const OpenshiftStreamsFederated: React.FunctionComponent<OpenshiftStreamsFederat
           }}
         >
           <ModalProvider>
-            <PaginationProvider>
-              <OpenshiftStreams
-                preCreateInstance={preCreateInstance}
-                tokenEndPointUrl={tokenEndPointUrl}
-              />
-            </PaginationProvider>
-            <KasModalLoader />
+            <InstanceDrawerContextProvider>
+              <PaginationProvider>
+                <KasLayout tokenEndPointUrl={tokenEndPointUrl}>
+                  <StreamsTableConnected
+                    preCreateInstance={preCreateInstance}
+                  />
+                </KasLayout>
+              </PaginationProvider>
+              <KasModalLoader />
+            </InstanceDrawerContextProvider>
           </ModalProvider>
         </FederatedContext.Provider>
       </I18nextProvider>
