+++ src/app/modules/OpenshiftStreams/components/KafkaEmptyState/index.ts
@@ -1,0 +1,1 @@
+export * from './KafkaEmptyState';
