+++ src/app/modules/OpenshiftStreams/components/KasLayout/index.ts
@@ -1,0 +1,1 @@
+export * from './KasLayout';
