+++ src/app/modules/OpenshiftStreams/components/StatusColumn/index.ts
@@ -1,0 +1,1 @@
+export * from './StatusColumn';
