+++ src/app/modules/OpenshiftStreams/components/StreamsTable/index.ts
@@ -1,0 +1,1 @@
+export * from './StreamsTable';
