+++ src/app/modules/OpenshiftStreams/components/StreamsTableConnected/StreamsTableConnected.css
