+++ src/app/modules/OpenshiftStreams/components/StreamsTableConnected/index.ts
@@ -1,0 +1,2 @@
+export * from './StreamsTableConnected';
+export { KafkaStatusAlerts } from '@app/modules/OpenshiftStreams/components/StreamsTableConnected/KafkaStatusAlerts';
