--- src/app/modules/OpenshiftStreams/components/StreamsTableView/index.ts
@@ -1,3 +1,0 @@
-export * from './StreamsTableView';
-export * from './StatusColumn';
-export * from './StreamsToolbar';
