+++ src/app/modules/OpenshiftStreams/components/StreamsToolbar/StreamsToolbar.css
