+++ src/app/modules/OpenshiftStreams/components/StreamsToolbar/index.ts
@@ -1,0 +1,1 @@
+export * from './StreamsToolbar';
