+++ src/app/modules/OpenshiftStreams/components/TableFilters/NameFilter.tsx
@@ -1,7 +1,7 @@
-import { FilterProps } from '@app/modules/OpenshiftStreams/components/StreamsTableView/Filters/types';
+import { FilterProps } from '@app/modules/OpenshiftStreams/components/TableFilters/types';
 import React, { useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useTooltipContent } from '@app/modules/OpenshiftStreams/components/StreamsTableView/Filters/hooks';
+import { useTooltipContent } from '@app/modules/OpenshiftStreams/components/TableFilters/hooks';
 import {
   Button,
   ButtonVariant,
