+++ src/app/modules/OpenshiftStreams/components/TableFilters/hooks.tsx
