+++ src/app/modules/OpenshiftStreams/components/TableFilters/index.ts
@@ -1,0 +1,3 @@
+export * from './StreamsFilterGroup';
+export { FilterType } from '@app/modules/OpenshiftStreams/components/TableFilters/types';
+export { FilterValue } from '@app/modules/OpenshiftStreams/components/TableFilters/types';
