+++ src/app/modules/OpenshiftStreams/components/TableFilters/types.ts
@@ -1,5 +1,4 @@
 import { ToolbarChip } from '@patternfly/react-core';
-import { FilterValue } from '@app/modules/OpenshiftStreams/components';
 import { KeyValueOptions } from '@app/utils';
 
 export type FilterProps = {
@@ -19,4 +18,14 @@ export type FilterProps = {
   ) => void;
   removeFilterValue: (value: string) => void;
   isDisabledSelectOption: (key: string, optionValue: string) => boolean;
+};
+
+export type FilterValue = {
+  value: string;
+  isExact: boolean;
+};
+
+export type FilterType = {
+  filterKey: string;
+  filterValue: FilterValue[];
 };
