+++ src/app/modules/OpenshiftStreams/components/Unauthorized/index.ts
@@ -1,0 +1,1 @@
+export * from './Unauthorized';
