+++ src/app/modules/OpenshiftStreams/components/index.ts
@@ -1,2 +1,7 @@
-export * from './StreamsTableView';
-export * from '../../InstanceDrawer';
+export * from './StreamsTable';
+export * from './StatusColumn';
+export * from './StreamsToolbar';
+export * from './TableFilters';
+export * from './Unauthorized';
+export * from './KafkaEmptyState';
+export * from './KasLayout';
