+++ src/app/modules/OpenshiftStreams/index.ts
@@ -1,1 +1,1 @@
-export * from './OpenshiftStreamsConnected';
+export * from './OpenshiftStreamsStandalone';
